dv-5 RSS408-219 - further support for integration testing
optional keystore.sha1 property to check the keystore loaded is the one we expect
logging of digests of IVs
logging of digests of AES keys
warning if encrypted chunks is not decrypted chunk size + 16
warning if decrypted chunks is not encrypted chunk size - 16
DecryptedTarBuilderTest - checks that encrypted chunks can be decrypted and put back into a tar file.
