+++ datavault-common/src/main/java/org/datavaultplatform/common/crypto/EncryptionValidator.java
@@ -11,6 +11,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.bouncycastle.util.Arrays;
+import org.datavaultplatform.common.storage.Verify;
 import org.datavaultplatform.common.task.Context.AESMode;
 import org.springframework.util.Assert;
 
@@ -23,7 +24,8 @@ public class EncryptionValidator {
   private static final String LABEL_KEY_FOR_PRIVATE_KEYS = "key for SSHPrivateKeys";
   private static final String LABEL_KEY_FOR_DATA = "key for Data";
 
-  public void validate(boolean validateDataKey, boolean validatePrivateKeysKey) {
+  public void validate(boolean validateDataKey, boolean validatePrivateKeysKey, String encryptionKeyStoreSHA1) {
+    validateKeyStoreSHA1(encryptionKeyStoreSHA1);
     if (Encryption.getVaultEnable() && Encryption.getKeystoreEnable()) {
       log.warn("The Encryption is setup with BOTH HashiCorpVault and Java Keystore");
     }
@@ -47,6 +49,29 @@ public class EncryptionValidator {
       validateEncryptionConfig(LABEL_KEY_FOR_PRIVATE_KEYS, null);
     }
     Encryption.logKeyDigests();
+  }
+
+  @SneakyThrows
+  private void validateKeyStoreSHA1(String expectedKeyStoreSha1) {
+    if (!Encryption.getKeystoreEnable()) {
+      log.warn("Encryption KeyStore is not enabled");
+      return;
+    }
+
+    String keyStorePath = Encryption.getKeystorePath();
+    File keyStoreFile = new File(keyStorePath);
+
+    Assert.isTrue(keyStoreFile.exists(), "KeyStore file does not exist: " + keyStorePath);
+    Assert.isTrue(keyStoreFile.canRead(), "KeyStore file is not readable: " + keyStorePath);
+
+    String actualKeyStoreSha1 = Verify.getDigest(keyStoreFile);
+    log.info("Encryption KeyStore [{}], actual SHA1[{}]", keyStorePath, actualKeyStoreSha1);
+    if (StringUtils.isBlank(expectedKeyStoreSha1)) {
+      return;
+    }
+    Assert.isTrue(expectedKeyStoreSha1.equalsIgnoreCase(actualKeyStoreSha1),
+          () -> String.format("KeyStore [%s]: actual SHA1[%s] does match expected SHA1[%s]",
+            keyStorePath, actualKeyStoreSha1, expectedKeyStoreSha1));
   }
 
   protected void validateKeyNames(boolean validateDataKey, boolean validatePrivateKeysKey) {
