+++ dv5/local-byodb/scripts/runLocalByodbBroker.sh
@@ -18,6 +18,7 @@ cd $PROJECT_ROOT
  SPRING_DATASOURCE_URL='jdbc:mysql://localhost:3306/test?useSSL=false&serverTimezone=UTC' \
 #KEYSTORE_ENABLE=true \
 #KEYSTORE_PATH=XXX/PATH/TO/KEYSTORE.ks \
+#KEYSTORE_SHA1=4499a0663669cc06e7099632ba360c15a404943c \
 #VAULT_PRIVATEKEYENCRYPTIONKEYNAME=XXX-KEYNAME-FOR-PRIVATE-KEYS \
 #VAULT_DATAENCRYPTIONKEYNAME=XXX-KEYNAME-FOR-DATA-KEYS \
 #KEYSTORE_PASSWORD=XXX-PASSWORD \
