+++ dv5/local-byodb/scripts/runLocalByodbWorker.sh
@@ -10,6 +10,7 @@ cd $PROJECT_ROOT
  DATAVAULT_HOME="$PROJECT_ROOT/dv5/local-byodb/props/worker" \
 #KEYSTORE_ENABLE=true \
 #KEYSTORE_PATH=XXX/PATH/TO/KEYSTORE.ks \
+#KEYSTORE_SHA1=4499a0663669cc06e7099632ba360c15a404943c \
 #VAULT_PRIVATEKEYENCRYPTIONKEYNAME=XXX-KEYNAME-FOR-PRIVATE-KEYS \
 #VAULT_DATAENCRYPTIONKEYNAME=XXX-KEYNAME-FOR-DATA-KEYS \
 #KEYSTORE_PASSWORD=XXX-PASSWORD \
