+++ dv5/local/scripts/runLocalWorker.sh
@@ -18,6 +18,7 @@ LOGFILE_SETTINGS=\
 #ENCRYPTION_SETTINGS=\
 #"KEYSTORE_ENABLE=true \
 # KEYSTORE_PATH=XXX/PATH/TO/KEYSTORE.ks \
+# KEYSTORE_SHA1=4499a0663669cc06e7099632ba360c15a404943c \
 # VAULT_PRIVATEKEYENCRYPTIONKEYNAME=XXX-KEYNAME-FOR-PRIVATE-KEYS \
 # VAULT_DATAENCRYPTIONKEYNAME=XXX-KEYNAME-FOR-DATA-KEYS \
 # KEYSTORE_PASSWORD=XXX-PASSWORD \
